+++ Projects/Core/Sources/Extension/UICollectionView+Extension.swift
@@ -8,6 +8,17 @@
 
 import UIKit
 
+public extension UICollectionView {
+    
+    func register<T>(_ cellClass: T.Type) where T: UICollectionViewCell {
+        self.register(cellClass, forCellWithReuseIdentifier: cellClass.defaultReuseIdentifier)
+    }
+    
+    func dequeueCell<T>(_ cellClass: T.Type, for indexPath: IndexPath) -> T? where T: UICollectionViewCell {
+        return self.dequeueReusableCell(withReuseIdentifier: cellClass.defaultReuseIdentifier, for: indexPath) as? T
+    }
+}
+
 public extension UICollectionViewCell {
     static var defaultReuseIdentifier: String {
         return String(describing: self)
