+++ Projects/Core/Sources/Extension/UITableViewCell+Extension.swift
@@ -8,6 +8,18 @@
 
 import UIKit
 
+public extension UITableView {
+    
+    func register<T>(_ cellClass: T.Type) where T: UITableViewCell {
+        self.register(cellClass.self, forCellReuseIdentifier: cellClass.defaultReuseIdentifier)
+    }
+    
+    func dequeueCell<T>(_ cellClass: T.Type, for indexPath: IndexPath) -> T? where T: UITableViewCell {
+        return self.dequeueReusableCell(withIdentifier: cellClass.defaultReuseIdentifier, for: indexPath) as? T
+    }
+    
+}
+
 public extension UITableViewCell {
     static var defaultReuseIdentifier: String {
         return String(describing: self)
